requirements-deploy.txt
# CIRCUITRON – Deployment requirements (Render)
# TrOCR / transformers excluded to keep image small and deploy fast.
# Only "fast" OCR mode (custom CRNN) is available in production.

# PyTorch CPU-only (much smaller than CUDA builds)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.134.0
filelock==3.24.3
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.1.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scipy==1.17.1
starlette==0.52.1
typing_extensions==4.15.0
ultralytics
urllib3==2.6.3
uvicorn==0.41.0
gunicorn==23.0.0
